+++ mopidy/audio/actor.py
@@ -435,6 +435,7 @@ class Audio(pykka.ThreadingActor):
         self._config = config
         self._target_state = Gst.State.NULL
         self._buffering = False
+        self._live_stream = False
         self._tags = {}
         self._pending_uri = None
         self._pending_tags = None
@@ -570,9 +571,13 @@ class Audio(pykka.ThreadingActor):
         else:
             self._appsrc.reset()
 
+        if self._live_stream and hasattr(source.props, "is_live"):
+            gst_logger.debug("Enabling live stream mode")
+            source.set_live(True)
+
         utils.setup_proxy(source, self._config["proxy"])
 
-    def set_uri(self, uri):
+    def set_uri(self, uri, live_stream=False):
         """
         Set URI of audio to be played.
 
@@ -580,6 +585,9 @@ class Audio(pykka.ThreadingActor):
 
         :param uri: the URI to play
         :type uri: string
+        :param live_stream: disables buffering, reducing latency for stream,
+            and discarding data when paused
+        :type live_stream: bool
         """
 
         # XXX: Hack to workaround issue on Mac OS X where volume level
@@ -591,6 +599,7 @@ class Audio(pykka.ThreadingActor):
 
         self._pending_uri = uri
         self._pending_tags = {}
+        self._live_stream = live_stream
         self._playbin.set_property("uri", uri)
 
         if self.mixer is not None and current_volume is not None: