+++ docs/changelog.rst
@@ -181,7 +181,9 @@ Tracklist controller
 Backend API
 -----------
 
-- (no changes yet)
+- Add :meth:`mopidy.backend.PlaybackProvider.is_live` which can be
+  re-implemented by Playback providers, to define if the custom
+  URI scheme is a live stream. (PR: :issue:`1845`)
 
 Models
 ------
@@ -263,6 +265,10 @@ Audio
 
 - Remove the method :meth:`mopidy.audio.Audio.emit_end_of_stream`, which has
   been deprecated since 1.0. (Fixes: :issue:`1465`, PR: :issue:`1705`)
+
+- Add live_stream option to the method :meth:`mopidy.audio.Audio.set_uri`,
+  which disables buffering, reducing latency for stream,
+  and discarding data when paused. (PR: :issue:`1845`)
 
 Internals
 ---------